# sdks/swagger/alpha.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "alpha.proto",
    "version": "version not set"
  },
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/alpha/player/capacity": {
      "get": {
        "summary": "Get the last player capacity that was set through the SDK.\nIf the player capacity is set from outside the SDK, use SDK.GameServer() instead.",
        "operationId": "GetPlayerCapacity",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/alphaCount"
            }
          }
        },
        "tags": [
          "SDK"
        ]
      },
      "post": {
        "summary": "Change the player capacity to a new value.",
        "operationId": "SetPlayerCapacity",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/alphaEmpty"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/alphaCount"
            }
          }
        ],
        "tags": [
          "SDK"
        ]
      }
    },
    "/alpha/player/connect": {
      "post": {
        "summary": "Call when a player has connected.",
        "operationId": "PlayerConnect",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/alphaEmpty"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/alphaPlayerId"
            }
          }
        ],
        "tags": [
          "SDK"
        ]
      }
    },
    "/alpha/player/count": {
      "get": {
        "summary": "get the current player count",
        "operationId": "GetPlayerCount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/alphaCount"
            }
          }
        },
        "tags": [
          "SDK"
        ]
      }
    },
    "/alpha/player/disconnect": {
      "post": {
        "summary": "Call when a player has disconnected.",
        "operationId": "PlayerDisconnect",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/alphaEmpty"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/alphaPlayerId"
            }
          }
        ],
        "tags": [
          "SDK"
        ]
      }
    }
  },
  "definitions": {
    "alphaCount": {
      "type": "object",
      "properties": {
        "count": {
          "type": "string",
          "format": "int64"
        }
      },
      "description": "Store a count variable."
    },
    "alphaEmpty": {
      "type": "object",
      "title": "I am Empty"
    },
    "alphaPlayerId": {
      "type": "object",
      "properties": {
        "playerId": {
          "type": "string"
        }
      },
      "description": "The unique identifier for a given player."
    }
  }
}